+++ src/sys/display/x11/event.rs
@@ -3,8 +3,9 @@
 //! Defines [`XEvent`].
 //
 
-use super::{KeyRepeatFilter, XkbState, raw};
+use super::{KeyRepeatFilter, XkbState, raw, xcb_event_code};
 use crate::{EventKey, EventTimestamp, KeyState, Libc, c_void};
+// use crate::{EventKind, EventWindow};
 
 /// Wrapper for an XCB event.
 #[derive(Debug)]
@@ -20,32 +21,67 @@ impl XEvent {
 
     /// Returns the event timestamp in backend-specific milliseconds, if present.
     pub fn timestamp(&self) -> Option<EventTimestamp> {
-        if self.is_key() {
-            let ev = self.raw as *const raw::xcb_key_press_event_t;
+        type XcbAnyInputEvent = raw::xcb_key_press_event_t;
+        if self.is_key() | self.is_button() | self.is_motion() | self.is_enter() | self.is_leave() {
+            let ev = self.raw as *const XcbAnyInputEvent;
             Some(EventTimestamp::from_millis_u32((unsafe { *ev }).time))
         } else {
             None
         }
     }
 
-    /// Returns true if this is an expose (repaint) event.
-    #[inline(always)]
-    pub fn is_expose(&self) -> bool {
-        self.response_type() == raw::xcb_event_code::XCB_EXPOSE as u8
+    // TODO
+    // pub fn event_kind(&self) -> EventKind {
+    //     match self.response_type() {
+    //         raw::XCB_EXPOSE => EventKind::Window(EventWindow::RedrawRequested),
+    //         _ => EventKind::None,
+    //     }
+    // }
+
+    /// Whether this is a key event.
+    pub fn is_key(&self) -> bool { self.is_key_press() || self.is_key_release() }
+
+    /// Whether this is a key-press event.
+    pub fn is_key_press(&self) -> bool {
+        self.response_type() == xcb_event_code::XCB_KEY_PRESS as u8
+    }
+    /// Whether this is a key-release event.
+    pub fn is_key_release(&self) -> bool {
+        self.response_type() == xcb_event_code::XCB_KEY_RELEASE as u8
     }
 
-    /// Returns true if this is a key event.
-    #[inline(always)]
-    pub fn is_key(&self) -> bool { self.is_key_press() || self.is_key_release() }
-    /// Returns true if this is a key-press event.
-    #[inline(always)]
-    pub fn is_key_press(&self) -> bool {
-        self.response_type() == raw::xcb_event_code::XCB_KEY_PRESS as u8
+    /// Whether this is a button event.
+    pub fn is_button(&self) -> bool { self.is_button_press() || self.is_button_release() }
+
+    /// Whether this is a button-press event.
+    pub fn is_button_press(&self) -> bool {
+        self.response_type() == xcb_event_code::XCB_BUTTON_PRESS as u8
     }
-    /// Returns true if this is a key-press event.
-    #[inline(always)]
-    pub fn is_key_release(&self) -> bool {
-        self.response_type() == raw::xcb_event_code::XCB_KEY_RELEASE as u8
+    /// Whether this is a button-release event.
+    pub fn is_button_release(&self) -> bool {
+        self.response_type() == xcb_event_code::XCB_BUTTON_RELEASE as u8
+    }
+
+    /// Whether this is a motion event.
+    pub fn is_motion(&self) -> bool {
+        self.response_type() == xcb_event_code::XCB_MOTION_NOTIFY as u8
+    }
+
+    /// Whether this is an enter event.
+    pub fn is_enter(&self) -> bool {
+        self.response_type() == xcb_event_code::XCB_ENTER_NOTIFY as u8
+    }
+    /// Whether this is a leave event.
+    pub fn is_leave(&self) -> bool {
+        self.response_type() == xcb_event_code::XCB_LEAVE_NOTIFY as u8
+    }
+    /// Whether this is an expose (repaint) event.
+    pub fn is_expose(&self) -> bool {
+        self.response_type() == xcb_event_code::XCB_EXPOSE as u8
+    }
+    /// Whether this is a client message.
+    pub fn is_client_message(&self) -> bool {
+        self.response_type() == xcb_event_code::XCB_CLIENT_MESSAGE as u8
     }
 
     /* internals */