+++ src/sys/display/x11/raw/xcb_values.rs
@@ -3,9 +3,11 @@
 //! Values from `xproto.h`
 //
 // TOC
-//  - xcb_event_code
-//  - xcb_error_code
-//  - xcb_request_code
+// - xcb_event_code
+// - xcb_error_code
+// - xcb_request_code
+// - xcb_prop_mode
+// - window attribute flags
 
 #![allow(non_camel_case_types)]
 
@@ -18,39 +20,87 @@
 /// where the low 7 bits select one of these opcodes.
 #[repr(u8)]
 pub(crate) enum xcb_event_code {
+    /// Key down
     XCB_KEY_PRESS = 2,
+    /// Key up
     XCB_KEY_RELEASE = 3,
+    /// Mouse button down
     XCB_BUTTON_PRESS = 4,
+    /// Mouse button up
     XCB_BUTTON_RELEASE = 5,
+    /// Mouse/pointer movement
     XCB_MOTION_NOTIFY = 6,
+    /// Pointer entered window
     XCB_ENTER_NOTIFY = 7,
+    /// Pointer left window
     XCB_LEAVE_NOTIFY = 8,
+    /// The window gained keyboard focus.
     XCB_FOCUS_IN = 9,
+    /// The window lost keyboard focus.
     XCB_FOCUS_OUT = 10,
+    /// Legacy "keyboard mapping changed"
     XCB_KEYMAP_NOTIFY = 11,
+    /// Window region became visible, redraw needed
     XCB_EXPOSE = 12,
+    /// Obscure, only for certain graphics ops.
     XCB_GRAPHICS_EXPOSURE = 13,
+    /// Confirms no Exposure occurred.
     XCB_NO_EXPOSURE = 14,
+    /// Visibility has changed.
     XCB_VISIBILITY_NOTIFY = 15,
+
+    /* window life cycle*/
+    /// A window was created.
     XCB_CREATE_NOTIFY = 16,
+    /// A window was destroyed.
     XCB_DESTROY_NOTIFY = 17,
+    /// Window hidden/minimized.
     XCB_UNMAP_NOTIFY = 18,
+    /// Window shown.
     XCB_MAP_NOTIFY = 19,
+
+    /* window reparenting */
+    /// Window manager asking to map.
     XCB_MAP_REQUEST = 20,
+    /// WM moved window into a frame (normal under tiling/stacking WMs).
     XCB_REPARENT_NOTIFY = 21,
+
+    /* geometry changes*/
+    /// Position/size changed (your resize event).
     XCB_CONFIGURE_NOTIFY = 22,
+    /// WM requesting configure.
     XCB_CONFIGURE_REQUEST = 23,
+    /// Rare offset change.
     XCB_GRAVITY_NOTIFY = 24,
+    /// Legacy; usually bypassed by modern WMs.
     XCB_RESIZE_REQUEST = 25,
+
+    /* stacking / circulation */
     XCB_CIRCULATE_NOTIFY = 26,
     XCB_CIRCULATE_REQUEST = 27,
+
+    /* properties & selection */
+    /// Fires for every change to any window property.
+    /// E.g. Title changed, WM hints changed, etc.
     XCB_PROPERTY_NOTIFY = 28,
+    /// Lost clipboard ownership.
     XCB_SELECTION_CLEAR = 29,
+    /// Another app wants clipboard contents.
     XCB_SELECTION_REQUEST = 30,
+    /// Clipboard transfer finished.
     XCB_SELECTION_NOTIFY = 31,
+
+    /* color map */
+    /// Obsolete. Color map change.
     XCB_COLORMAP_NOTIFY = 32,
+
+    /* client messages */
+    /// Used for WM_DELETE_WINDOW (close button), _NET_WM_STATE (fullscreen, maximized, minimized).
     XCB_CLIENT_MESSAGE = 33,
+    /// Obsolete. XT keyboard remap.
     XCB_MAPPING_NOTIFY = 34,
+
+    /// Generic events. For XInput2 (high-resolution mouse, raw input, multi-touch).
     XCB_GE_GENERIC = 35,
 }
 
@@ -207,6 +257,83 @@ enum xcb_request_code {
     XCB_GET_MODIFIER_MAPPING = 119,
     //
     XCB_NO_OPERATION = 127,
+}
+
+///
+#[repr(u8)]
+pub(crate) enum xcb_atom_enum_t {
+    XCB_ATOM_ANY = 0,
+    XCB_ATOM_PRIMARY = 1,
+    XCB_ATOM_SECONDARY = 2,
+    XCB_ATOM_ARC = 3,
+    XCB_ATOM_ATOM = 4,
+    XCB_ATOM_BITMAP = 5,
+    XCB_ATOM_CARDINAL = 6,
+    XCB_ATOM_COLORMAP = 7,
+    XCB_ATOM_CURSOR = 8,
+    XCB_ATOM_CUT_BUFFER0 = 9,
+    XCB_ATOM_CUT_BUFFER1 = 10,
+    XCB_ATOM_CUT_BUFFER2 = 11,
+    XCB_ATOM_CUT_BUFFER3 = 12,
+    XCB_ATOM_CUT_BUFFER4 = 13,
+    XCB_ATOM_CUT_BUFFER5 = 14,
+    XCB_ATOM_CUT_BUFFER6 = 15,
+    XCB_ATOM_CUT_BUFFER7 = 16,
+    XCB_ATOM_DRAWABLE = 17,
+    XCB_ATOM_FONT = 18,
+    XCB_ATOM_INTEGER = 19,
+    XCB_ATOM_PIXMAP = 20,
+    XCB_ATOM_POINT = 21,
+    XCB_ATOM_RECTANGLE = 22,
+    XCB_ATOM_RESOURCE_MANAGER = 23,
+    XCB_ATOM_RGB_COLOR_MAP = 24,
+    XCB_ATOM_RGB_BEST_MAP = 25,
+    XCB_ATOM_RGB_BLUE_MAP = 26,
+    XCB_ATOM_RGB_DEFAULT_MAP = 27,
+    XCB_ATOM_RGB_GRAY_MAP = 28,
+    XCB_ATOM_RGB_GREEN_MAP = 29,
+    XCB_ATOM_RGB_RED_MAP = 30,
+    XCB_ATOM_STRING = 31,
+    XCB_ATOM_VISUALID = 32,
+    XCB_ATOM_WINDOW = 33,
+    XCB_ATOM_WM_COMMAND = 34,
+    XCB_ATOM_WM_HINTS = 35,
+    XCB_ATOM_WM_CLIENT_MACHINE = 36,
+    XCB_ATOM_WM_ICON_NAME = 37,
+    XCB_ATOM_WM_ICON_SIZE = 38,
+    XCB_ATOM_WM_NAME = 39,
+    XCB_ATOM_WM_NORMAL_HINTS = 40,
+    XCB_ATOM_WM_SIZE_HINTS = 41,
+    XCB_ATOM_WM_ZOOM_HINTS = 42,
+    XCB_ATOM_MIN_SPACE = 43,
+    XCB_ATOM_NORM_SPACE = 44,
+    XCB_ATOM_MAX_SPACE = 45,
+    XCB_ATOM_END_SPACE = 46,
+    XCB_ATOM_SUPERSCRIPT_X = 47,
+    XCB_ATOM_SUPERSCRIPT_Y = 48,
+    XCB_ATOM_SUBSCRIPT_X = 49,
+    XCB_ATOM_SUBSCRIPT_Y = 50,
+    XCB_ATOM_UNDERLINE_POSITION = 51,
+    XCB_ATOM_UNDERLINE_THICKNESS = 52,
+    XCB_ATOM_STRIKEOUT_ASCENT = 53,
+    XCB_ATOM_STRIKEOUT_DESCENT = 54,
+    XCB_ATOM_ITALIC_ANGLE = 55,
+    XCB_ATOM_X_HEIGHT = 56,
+    XCB_ATOM_QUAD_WIDTH = 57,
+    XCB_ATOM_WEIGHT = 58,
+    XCB_ATOM_POINT_SIZE = 59,
+    XCB_ATOM_RESOLUTION = 60,
+    XCB_ATOM_COPYRIGHT = 61,
+    XCB_ATOM_NOTICE = 62,
+    XCB_ATOM_FONT_NAME = 63,
+    XCB_ATOM_FAMILY_NAME = 64,
+    XCB_ATOM_FULL_NAME = 65,
+    XCB_ATOM_CAP_HEIGHT = 66,
+    XCB_ATOM_WM_CLASS = 67,
+    XCB_ATOM_WM_TRANSIENT_FOR = 68,
+}
+impl xcb_atom_enum_t {
+    pub const XCB_ATOM_NONE: Self = Self::XCB_ATOM_ANY;
 }
 
 /* property modes */