+++ src/sys/display/x11/window.rs
@@ -6,7 +6,7 @@
 #![allow(unused)]
 
 use super::raw;
-use crate::{Extent, Libc, Position, XDisplay, XError, XEvent, lets};
+use crate::{Extent, Libc, Position, XDisplay, XError, XEvent, lets, x11_intern_atom};
 
 /// X11 window created through XCB.
 ///
@@ -31,13 +31,35 @@ impl XWindow {
         let win: u32 = unsafe { raw::xcb_generate_id(conn) }; // generate window ID
         let screen = unsafe { &*display.screen }; // extract screen
 
-        // create window
-        let values: [u32; 2] = [screen.white_pixel,
+        // create window, supported events:
+        let values: [u32; 2] = [screen.black_pixel,
+            // key input
             raw::XCB_EVENT_MASK_KEY_PRESS
             | raw::XCB_EVENT_MASK_KEY_RELEASE
+            // mouse button input
+            | raw::XCB_EVENT_MASK_BUTTON_PRESS
+            | raw::XCB_EVENT_MASK_BUTTON_RELEASE
+            | raw::XCB_EVENT_MASK_POINTER_MOTION
+            | raw::XCB_EVENT_MASK_BUTTON_1_MOTION
+            | raw::XCB_EVENT_MASK_BUTTON_2_MOTION
+            | raw::XCB_EVENT_MASK_BUTTON_3_MOTION
+            | raw::XCB_EVENT_MASK_BUTTON_4_MOTION
+            | raw::XCB_EVENT_MASK_BUTTON_5_MOTION
+            // | raw::XCB_EVENT_MASK_BUTTON_MOTION
+            // pointer enter/leave
+            | raw::XCB_EVENT_MASK_ENTER_WINDOW
+            | raw::XCB_EVENT_MASK_LEAVE_WINDOW
+            // focus in/out
+            | raw::XCB_EVENT_MASK_FOCUS_CHANGE
+            // expose
             | raw::XCB_EVENT_MASK_EXPOSURE
-            | raw::XCB_EVENT_MASK_KEY_PRESS
+            // | raw::XCB_EVENT_MASK_VISIBILITY_CHANGE
             | raw::XCB_EVENT_MASK_STRUCTURE_NOTIFY
+            // resize
+            // | raw::XCB_EVENT_MASK_RESIZE_REDIRECT
+            // | raw::XCB_EVENT_MASK_SUBSTRUCTURE_NOTIFY
+            // | raw::XCB_EVENT_MASK_SUBSTRUCTURE_REDIRECT
+            // | raw::XCB_EVENT_MASK_PROPERTY_CHANGE
         ];
         let mask: u32 = raw::XCB_CW_BACK_PIXEL | raw::XCB_CW_EVENT_MASK;
         unsafe {
@@ -50,6 +72,14 @@ impl XWindow {
         let gc_values: [u32; 1] = [screen.black_pixel];
         unsafe { raw::xcb_create_gc(conn, gc, win, raw::XCB_GC_FOREGROUND, gc_values.as_ptr()); }
         unsafe { raw::xcb_map_window(conn, win); } // show window
+
+        // enable WM_DELETE_WINDOW so the window manager can request closure
+        let wm_protocols = x11_intern_atom(conn, b"WM_PROTOCOLS");
+        let wm_delete_window = x11_intern_atom(conn, b"WM_DELETE_WINDOW");
+        let data = wm_delete_window.to_ne_bytes();
+        unsafe { raw::xcb_change_property(conn, raw::xcb_prop_mode::XCB_PROP_MODE_REPLACE as u8,
+            win, wm_protocols, raw::xcb_atom_enum_t::XCB_ATOM_ATOM as u32, 32, 1, data.as_ptr()); }
+
         let window = Self { display: conn, win, gc, width, height, depth: display.depth };
         window.flush();
         Ok(window)